api/requirements.txt
numpy==1.26.4
imageio==2.37.0
imageio-ffmpeg==0.6.0
fastapi==0.119.0
uvicorn==0.37.0
huggingface_hub
pillow
psutil
torch
torchvision